Zipped ESRD Sketch Library
I created a new pull request because the previous was an unzipped file. Within this file, there is the ESRD Sketch Library which is intended to reduce redesign and foster visual consistency in the UI. Within this sketch library, users have Styles (Typography, Colors, and Iconolography) , Components, and Patterns (Page Templates) to reference from and pull as a symbol. The file itself can be referenced in any sketch file by adding the ESRD Sketch Library file as a library.